Rename SharedPageLayout to CasesLayout
