+++ apps/judicial-system/web/pages/fangelsi/index.ts
@@ -1,0 +1,3 @@
+import PrisonCases from '@island.is/judicial-system-web/src/routes/Shared/Cases/PrisonCases'
+
+export default PrisonCases
