+++ apps/judicial-system/web/src/components/index.ts
@@ -97,7 +97,7 @@ export { default as CaseResentExplanation } from './CaseResentExplanation/CaseRe
 export { AlertBanner } from './AlertBanner'
 export { default as RulingDateLabel } from './DateLabel/RulingDateLabel'
 export { default as DateLabel } from './DateLabel/DateLabel'
-export { default as SharedPageLayout } from './SharedPageLayout/SharedPageLayout'
+export { default as CasesLayout } from './CasesLayout'
 export { default as ReopenModal } from './ReopenModal/ReopenModal'
 export { default as RulingModifiedModal } from './RulingModifiedModal/RulingModifiedModal'
 export { default as CommentsInput } from './CommentsInput/CommentsInput'
