+++ apps/judicial-system/web/src/routes/PublicProsecutor/Cases/PublicProsecutorCases.tsx
@@ -7,7 +7,7 @@ import { errors, titles } from '@island.is/judicial-system-web/messages'
 import {
   Logo,
   PageHeader,
-  SharedPageLayout,
+  CasesLayout,
 } from '@island.is/judicial-system-web/src/components'
 import {
   CaseIndictmentRulingDecision,
@@ -56,7 +56,7 @@ export const PublicProsecutorCases: FC = () => {
   }, [resCases])
 
   return (
-    <SharedPageLayout>
+    <CasesLayout>
       <PageHeader title={formatMessage(titles.shared.cases)} />
       <div className={styles.logoContainer}>
         <Logo />
@@ -75,7 +75,7 @@ export const PublicProsecutorCases: FC = () => {
           <CasesReviewComplete cases={reviewedCases} loading={loading} />
         </>
       )}
-    </SharedPageLayout>
+    </CasesLayout>
   )
 }
