+++ apps/judicial-system/web/src/routes/Shared/Cases/PrisonCases.tsx
@@ -18,7 +18,7 @@ import {
   Logo,
   PageHeader,
   SectionHeading,
-  SharedPageLayout,
+  CasesLayout,
   TagAppealState,
   TagCaseState,
   UserContext,
@@ -285,7 +285,7 @@ export const PrisonCases: FC = () => {
   }
 
   return (
-    <SharedPageLayout>
+    <CasesLayout>
       <PageHeader title={formatMessage(titles.shared.cases)} />
       <div className={styles.logoContainer}>
         <Logo />
@@ -345,7 +345,7 @@ export const PrisonCases: FC = () => {
             : renderAlertMessage()}
         </>
       )}
-    </SharedPageLayout>
+    </CasesLayout>
   )
 }
